// src/ts/types.ts
export interface Nutrition {
    foodName: string;
    category: string;
    servingSize: string;
    energy: number;
    moisture: number;
    protein: number;
    fat: number;
    carbohydrate: number;
    sugar: number;
    fiber: number;
    calcium: number;
    potassium: number;
    sodium: number;
    vitaminD: number;
    cholesterol: number;
    saturatedFat: number;
    transFat: number;
  }